fix(canon): preserve symlinked incremental inputs
